requirements.txt
ccf==5.0.6
cryptography==43.0.1
detect-secrets==1.5.0
docker==7.1.0
GitPython==3.1.43
httpx==0.27.2
jinja2==3.1.4
matplotlib==3.7.5
pandas==2.0.3
paramiko==3.5.0
pyasn1==0.6.1
PyJWT==2.9.0
pre-commit==3.5.0
setuptools==72.0.0
pytest==8.3.3
python-dotenv==1.0.1